backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
pydantic==2.5.0
Pillow==10.1.0
pytesseract==0.3.10
python-dateutil==2.8.2
PyPDF2==3.0.1